+++ packages/types/src/documents/workspace/delegatedAuthCredential.ts
