+++ packages/types/src/documents/workspace/index.ts
@@ -3,6 +3,7 @@ export * from "./automation"
 export * from "./backup"
 export * from "./component"
 export * from "./datasource"
+export * from "./delegatedAuthCredential"
 export * from "./deployment"
 export * from "./layout"
 export * from "./links"
@@ -13,7 +14,6 @@ export * from "./role"
 export * from "./row"
 export * from "./rowAction"
 export * from "./screen"
-export * from "./sharepointCredential"
 export * from "./snippet"
 export * from "./sqlite"
 export * from "./table"
@@ -21,7 +21,7 @@ export * from "./theme"
 export * from "./user"
 export * from "./view"
 export * from "./webhook"
+export * from "./webSearchConfig"
 export * from "./workspace"
 export * from "./workspaceApp"
 export * from "./workspaceFavourite"
-export * from "./webSearchConfig"
